# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Festival API",
    "version": "1.0.0",
    "description": "API documentation for Festival management"
  },
  "paths": {
    "/api/events": {
      "get": {
        "summary": "Get all events",
        "description": "Retrieve a list of all events.",
        "responses": {
          "200": {
            "description": "A list of events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string" },
                      "name": { "type": "string" },
                      "date": { "type": "string", "format": "date" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new event",
        "description": "Add a new event to the system.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "date": { "type": "string", "format": "date" }
                },
                "required": ["name", "date"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created successfully"
          }
        }
      }
    },
...
  }
}